Sat, Oct 29, 2022, 3:42 PM +00:00

+++ volume/volume.c
@@ -49,7 +49,7 @@ int main(int argc, char *argv[])
 
     // TODO: Read samples from input file and write updated data to output file
     uint16_t buffer;
-    while (fread(&buffer, sizeof(uint), 1, input))
+    while (fread(&buffer, sizeof(uint16_t), 1, input))
     {
         buffer *= factor;
         fwrite(&buffer, 2, 1, output);
